+++ source/extensions/omni.isaac.articulation_inspector/omni/isaac/articulation_inspector/extension.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2018-2021, NVIDIA CORPORATION.  All rights reserved.
+# Copyright (c) 2018-2023, NVIDIA CORPORATION.  All rights reserved.
 #
 # NVIDIA CORPORATION and its licensors retain all intellectual property
 # and proprietary rights in and to this software, related documentation
@@ -18,6 +18,7 @@ import omni.ui as ui
 import omni.usd
 import omni.timeline
 from omni.kit.menu.utils import add_menu_items, remove_menu_items, MenuItemDescription
+from omni.isaac.ui.menu import make_menu_item_description
 from omni.isaac.core.utils.stage import get_stage_units
 from omni.isaac.core.utils.prims import get_prim_object_type
 from omni.isaac.core.articulations import Articulation
@@ -62,7 +63,7 @@ class Extension(omni.ext.IExt):
         self._models = {}
         self._ext_id = ext_id
         menu_items = [
-            MenuItemDescription(name=EXTENSION_NAME, onclick_fn=lambda a=weakref.proxy(self): a._menu_callback())
+            make_menu_item_description(ext_id, EXTENSION_NAME, lambda a=weakref.proxy(self): a._menu_callback())
         ]
         self._menu_items = [MenuItemDescription(name="Workflows", sub_menu=menu_items)]
         add_menu_items(self._menu_items, "Isaac Utils")