+++ source/extensions/omni.isaac.demos/omni/isaac/demos/navigation_preview.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2018-2021, NVIDIA CORPORATION.  All rights reserved.
+# Copyright (c) 2018-2023, NVIDIA CORPORATION.  All rights reserved.
 #
 # NVIDIA CORPORATION and its licensors retain all intellectual property
 # and proprietary rights in and to this software, related documentation
@@ -13,6 +13,7 @@ import omni.ext
 import omni.ui as ui
 import omni.physx as _physx
 from omni.kit.menu.utils import add_menu_items, remove_menu_items, MenuItemDescription
+from omni.isaac.ui.menu import make_menu_item_description
 from pxr import UsdGeom
 import math
 from omni.isaac.ui.ui_utils import (
@@ -74,7 +75,7 @@ class Extension(omni.ext.IExt):
         # self._window.set_visibility_changed_fn(self._on_window)
 
         menu_items = [
-            MenuItemDescription(name=EXTENSION_NAME, onclick_fn=lambda a=weakref.proxy(self): a._menu_callback())
+            make_menu_item_description(ext_id, EXTENSION_NAME, lambda a=weakref.proxy(self): a._menu_callback())
         ]
         self._menu_items = [MenuItemDescription(name="Demos", sub_menu=menu_items)]
         add_menu_items(self._menu_items, "Isaac Examples")