+++ source/extensions/omni.isaac.demos/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.3.2] - 2023-01-06
+### Fixed
+- onclick_fn warning when creating UI
+
 ## [0.3.1] - 2022-09-07
 ### Fixed
 - Fixes for kit 103.5